refactor(document processor): Move doc extraction and cleaning functions
Move to document extraction and cleaning functions to
cl.lib.document_processors & related imports
